api/requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.6.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==2.4.0
channels-redis==3.3.0
charset-normalizer==2.0.12
cloudinary==1.29.0
constantly==15.1.0
cryptography==37.0.4
daphne==2.5.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.12.0
django-environ==0.8.1
django-filter==21.1
django-webpack-loader==1.0.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
environ==1.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
Pillow==9.1.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.9
uWSGI==2.0.20
whitenoise==5.2.0
zope.interface==5.4.0
